pyproject.toml
[tool.poetry]
name = "smart_bear"
version = "0.1.0"
description = ""
authors = ["Shawn Koh <shawn@shawnkoh.sg>"]

[tool.poetry.scripts]
smart-bear = "smart_bear.cli:app"
safe-bear = "smart_bear.safecli:app"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
beautifulsoup4 = "^4.10.0"
lxml = "^4.6.3"
regex = "^2021.10.23"
unmarkd = "^0.1.7"
psutil = "^5.8.0"
python-dotenv = "^0.19.1"
typer = {extras = ["all"], version = "^0.4.0"}
arrow = "^1.2.1"
anki = "^2.1.49"
furl = "^2.1.3"
parsy = { git="https://github.com/python-parsy/parsy", rev = "master" }
aqt = {extras = ["qt6"], version = "^2.1.49", allow-prereleases = true}
attrs = "^21.4.0"
more-itertools = "^8.12.0"
tqdm = "^4.62.3"
pyfunctional = {git = "https://github.com/EntilZha/PyFunctional", rev = "master"}
markdown2 = "^2.4.2"
pyperclip = "^1.8.2"
textualize = "^0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
hypothesis = "^6.24.0"
black = "^22.6"
flake8 = "^4.0.1"
pytest-clarity = "^1.0.1"
isort = "^5.10.1"
rich = "^11.2.0"
pycln = "^1.2.4"

[tool.black]
target-version = ["py310"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
